Rename sort to order
