+++ README.md
@@ -29,6 +29,15 @@ class Plant < ActiveRecord::Base
 end
 ```
 
+Validate the elements are ordered:
+
+```ruby
+class Plant < ActiveRecord::Base
+  validates :intake_times, array: { order: :asc }
+  validates :quantity, array: { order: :desc }
+end
+```
+
 ## Installation
 
 Add this line to your application's Gemfile:
@@ -44,7 +53,6 @@ And then execute:
 
 ## TODO
 
-- array elements are sorted
 - length of the array
 
 ## Development
