+++ spec/array_validator/order_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe 'array_validator#sort' do
+describe 'array_validator#order' do
   class Fluid
     include ActiveModel::Validations
     attr_accessor :intake_times
@@ -9,7 +9,7 @@ describe 'array_validator#sort' do
   let(:fluid) { Fluid.new }
 
   context 'ascending' do
-    before { fluid.class_eval { validates :intake_times, array: { sorted: :asc } } }
+    before { fluid.class_eval { validates :intake_times, array: { order: :asc } } }
 
     it 'returns true for valid order' do
       fluid.intake_times = [2, 4, 6, 8]
@@ -19,12 +19,12 @@ describe 'array_validator#sort' do
     it 'returns false for invalid order' do
       fluid.intake_times = [2, 6, 4, 8]
       expect(fluid.valid?).to eq(false)
-      expect(fluid.errors.full_messages.first).to eq 'Intake times is not sorted in ascending order'
+      expect(fluid.errors.full_messages.first).to eq 'Intake times is not ordered in ascending order'
     end
   end
 
   context 'descending' do
-    before { fluid.class_eval { validates :intake_times, array: { sorted: :desc } } }
+    before { fluid.class_eval { validates :intake_times, array: { order: :desc } } }
 
     it 'returns true for valid order' do
       fluid.intake_times = [8, 6, 4, 2]
@@ -34,16 +34,16 @@ describe 'array_validator#sort' do
     it 'returns false for invalid order' do
       fluid.intake_times = [8, 4, 6, 2]
       expect(fluid.valid?).to eq(false)
-      expect(fluid.errors.full_messages.first).to eq 'Intake times is not sorted in descending order'
+      expect(fluid.errors.full_messages.first).to eq 'Intake times is not ordered in descending order'
     end
   end
 
   context 'unsupported order' do
-    before { fluid.class_eval { validates :intake_times, array: { sorted: :jibberish } } }
+    before { fluid.class_eval { validates :intake_times, array: { order: :jibberish } } }
 
     it 'returns true for valid order' do
       fluid.intake_times = [8, 6, 4, 2]
-      expect { fluid.valid? }.to raise_error(ArgumentError, 'Not a supported sorting option')
+      expect { fluid.valid? }.to raise_error(ArgumentError, 'Not a supported order option')
     end
   end
 end
